requirements.txt
attributedict
celery
channels
channels_redis
django<3
django-allauth
django-invitations
django-ipware
django-model-utils
django-pdb
django-qr-code
django-redis
djangorestframework
django-rest-auth
django_smtp_ssl
django-money
ethereum
eth-abi<2
ethtoken
factory_boy
psycopg2-binary
pycoingecko
pycryptodome
pyjwt
pytest-asyncio
pytest-django
pytest-env
remote-pdb
raiden_contracts
uvicorn
